fix bugs of models in train.py, rm files related to the old layer.py

+++ src/core/scheduler/scheduler.cc
@@ -125,7 +125,7 @@ void Graph::Debug() {
     max_free_num = std::max(max_free_num, it.size());
   }
 
-  for (int i = std::max(nodes_.size(), blocks_.size()); i > 0; i /= 10, ++w) {
+  for (size_t i = std::max(nodes_.size(), blocks_.size()); i > 0; i /= 10, ++w) {
   }
 
   std::stringstream ss;
